pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "aibox-nlp"
version = "0.0.0"
description = "AiBox Natural Language Processing Toolkit."
readme = "README.md"
classifiers = ["Python :: 3.10", "Python :: 3.11"]
dependencies = [
    "scikit-learn~=1.2.2",
    "torch~=2.0.1",
    "numpy~=1.24.4",
    "pandas~=2.0.3",
    "platformdirs>=3.9.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
trees = ["lightgbm~=4.0.0", "catboost~=1.2", "xgboost~=1.7.6"]
embeddings = ["fasttext-wheel~=0.9.2"]
BR = [
    "pyspellchecker~=0.7.2",
    "spacy~=3.2.6",
    "transformers~=4.31.0",
    "sentence-transformers~=2.2.2",
    "cogroo4py~=0.5.1",
    "language-tool-python~=2.7.1",
    "python-Levenshtein~=0.21.1",
    "polyfuzz~=0.4.0",
    "lexical-diversity~=0.1.1",
    "fuzzywuzzy~=0.18.0",
    "fuzzysearch~=0.7.3",
    "gensim~=4.3.1",
    "cohmetrixBR~=0.0.7",
    "py-rouge~=1.1",
    "TRUNAJOD~=0.1.1",
    "unidecode~=1.3.6",
    "scipy~=1.10.1",
    "kenlm~=0.1",
    "gensim~=4.3.2",
    "nltk~=3.8.1",
]
all = ["aibox-nlp[BR]", "aibox-nlp[trees]", "aibox-nlp[embeddings]"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["aibox*"]
namespaces = true
